--- lognorm/tests/__init__.py
@@ -1,3 +1,0 @@
-"""
-Unit tests for :py:mod:`lognorm`.
-"""
